[12/16/25] CCV deployment code update (#1489)
* Add AdvancedPoolHooks bindings

* WIP - resolve compile issues

* Update deployment pkg

* Add transferFeeConfig

* Updates

* Update OnRamp optimizations

* WIP

* Imports

* Fix tests

* More tests

* Add CCVs above thresh to deployment

* CCVs above threshold to EVM

* Bump

* Decimals

* Precommit

* Bump

+++ chains/evm/contracts/test/helpers/OnRampHelper.sol
@@ -81,7 +81,7 @@ contract OnRampHelper is OnRamp {
 
   function validateDestChainAddress(
     bytes memory rawAddress,
-    uint8 addressBytesLength
+    uint256 addressBytesLength
   ) external pure returns (bytes memory) {
     return _validateDestChainAddress(rawAddress, addressBytesLength);
   }
